Use BigDecimal, not Double or Integer, to hold the value of numeric LiteralNode. Deprecate corresponding LiteralNode.create methods.
git-svn-id: https://olap4j.svn.sourceforge.net/svnroot/olap4j/trunk@381 c6a108a4-781c-0410-a6c6-c2d559e19af0

+++ src/org/olap4j/mdx/parser/impl/Scanner.java
@@ -297,7 +297,7 @@ class Scanner {
      * @return number literal token
      */
     private Symbol makeNumber(BigDecimal mantissa, int exponent) {
-        double d = mantissa.movePointRight(exponent).doubleValue();
+        BigDecimal d = mantissa.movePointRight(exponent);
         return makeSymbol(DefaultMdxParserSym.NUMBER, d);
     }
